feat: init UI
YOu are an expert in building Angular web applications, using the latest best practices to build small, simple applications that are fast and maintainable.

Read sample.html which is the html of a page of a website. I want you to use the colors, spacing, logo from this sample site to setup a standard set of styles in the global styles.scss that can be used across the rest of the site. Don't create custom classes unless needed, prefer styling native h1, h2, h3, a, a:visited, a:link, nav, footer, button, elements.

I want you to create a new angular application that has a single data.json file that powers the entire site Load it in a service and make it available to all components. This JSON will contain a nested list of ecosystems, packages in that ecosystem, versions in that package, and vulnerabilities in that version. It will look like this:

[{ecosystem: 'javascript', packages: [{name: 'esbuild', versions: [{version: "0.21.5", status: "end-of-life",heroDevsSupported: false, vulnerabilities: [{cve: 'CVE-202501-23423', severity: 'Medium'}]}]}]}]

Create the initial data.json from the list in the given HTML based on the column headers. Generate sample data where information is missing.

Then create three components wired up using the router at /, /:ecosystem/, and /:ecosystem:/:package:/. A home component with a list of ecosystems, and for each ecosystem show the number of vulnerabilities. The ecosystems should include: npm, maven, ruby, pip, apt, snapd, flatpak, and any other software package ecosystems.

When the user clicks on an ecosystem, take them to the view-ecosystem component which shows a list of packages in the ecosystem as a table. The user should be able to sort and search the list of packages. When the user clicks on a package, they sould be taken to the view-package page.

The view package page shows general information about the ecosystem at the top, then a table of versions, showing support status, release date, and vulnerability count by severity (critical, high, medium, low).

On any package on any page that is supported by herodevs, show a mini version of the herodevs logo (20px high) that links to herodevs.com.